Move classes that are services into Service
A little bit of namespace confusion here, Classes namespace is really
for plain objects, it makes more sense for these to be in the Service
namespace
