+++ src/Ilios/CliBundle/Command/ValidateLearningMaterialPathsCommand.php
@@ -9,7 +9,7 @@ use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Helper\Table;
 
 use Ilios\CoreBundle\Entity\Manager\LearningMaterialManager;
-use Ilios\CoreBundle\Classes\IliosFileSystem;
+use Ilios\CoreBundle\Service\IliosFileSystem;
 
 /**
  * Sync a user with their directory information
